feat(seed-admin): create admin user seeding script
